Column Selector
The column selector should not include the checkbox as that is always required and should not be able to be removed.

The script assumes that the first column is the checkbox but in this view there was an extra empty column.

This PR removes the extra column. I assume it is safe to do that as the contents of the column were commented out.

Signed-off-by: BrianTeeman <brian@teeman.net>